// font_end_khachhang/src/types/product.type.ts
export interface IVoucher {
  code: string;
  id: number;
  index: number;
  minBillValue: number;
  maxBillValue: number;
  name: string;
  percentReduce: number;
  quantity: number;
  status: number;
  type: boolean;
  startDate: string;
  endDate: string;
  customer: null | string;
}